refactor: Gamer.class 를 삭제하고 Dealer가 Player를 상속받도록 구조변경
